some basic programing i made practiced here

+++ Primitive Data type/src/primitive/dataTyppe/SolveAnEquation.java
@@ -5,7 +5,7 @@ public class SolveAnEquation {
         int a =25;
         float b =42.2323f;
 
-        double value = a*a + 2*(a*b) + b*b;
+        int value = (int)(a*a + 2*(a*b) + b*b);
         System.out.println("Value of the equaltion is : "+ value);
     }
 }
